use same page for project tabs
