+++ frontend/src/components/ProjectSettings.module.scss
@@ -1,8 +1,4 @@
 .container {
-    max-width: 50em;
-    padding: 1em;
-    margin: 0 auto;
-
     display: flex;
     flex-direction: column;
     gap: 1.5em;
