+++ frontend/src/pages/projects/[...project].module.scss
